requirements.txt
fastapi
uvicorn
vk_api
python-telegram-bot
deta
schedule
pillow